Refactor vehicle library validation boundaries
Split vehicle configuration shard expansion, typed row projection, and vehicle library validation rules into focused adapter-internal modules while preserving the public loading and validation facades.

Add focused regression coverage for gear ordering, tire plausibility, and source evidence resolution failures.
