+++ apps/server/tests/adapters/persistence/test_car_library_source_evidence.py
@@ -3,11 +3,16 @@
 from __future__ import annotations
 
 import json
+from dataclasses import replace
 from pathlib import Path
 
 import pytest
 
-from vibesensor.adapters.persistence.car_library_source_evidence import load_car_source_registry
+from vibesensor.adapters.persistence.car_library_source_evidence import (
+    load_car_source_registry,
+    validate_vehicle_configuration_source_evidence,
+)
+from vibesensor.adapters.persistence.vehicle_configurations import load_vehicle_configurations
 
 
 def test_load_car_source_registry_rejects_bad_source_pack_reference(tmp_path: Path) -> None:
@@ -33,3 +38,42 @@ def test_load_car_source_registry_rejects_bad_source_pack_reference(tmp_path: Pa
 
     with pytest.raises(ValueError, match="must start with"):
         load_car_source_registry(source_packs_dir=source_dir)
+
+
+def test_validate_vehicle_source_evidence_requires_refs_for_official_metadata() -> None:
+    config = load_vehicle_configurations()[0]
+    broken = replace(
+        config,
+        drivetrain_metadata=replace(
+            config.drivetrain_metadata,
+            confidence="official_exact",
+            evidence_refs=(),
+        ),
+    )
+
+    issues = validate_vehicle_configuration_source_evidence(
+        [broken],
+        registry=load_car_source_registry(),
+    )
+
+    assert [issue.rule for issue in issues] == ["missing_required_evidence_refs"]
+    assert "drivetrain" in issues[0].message
+
+
+def test_validate_vehicle_configuration_source_evidence_rejects_unknown_ref() -> None:
+    config = load_vehicle_configurations()[0]
+    broken = replace(
+        config,
+        drivetrain_metadata=replace(
+            config.drivetrain_metadata,
+            evidence_refs=("missing_pack:missing-source",),
+        ),
+    )
+
+    issues = validate_vehicle_configuration_source_evidence(
+        [broken],
+        registry=load_car_source_registry(),
+    )
+
+    assert [issue.rule for issue in issues] == ["missing_source_reference"]
+    assert "missing_pack:missing-source" in issues[0].message
